Use discriminated union type for Actions 🤯
this is so cool! 🎉 👯 🎊
this means we can get rid of the null check as the TS compiler can work
out that if were in an uncover or toggle action there will be coords,
it's also ensure when we use either of those actions we will be warned
if we haven't passed coords! 🤯 so good!